+++ Runtime/Api/UGC.cs
@@ -750,8 +750,31 @@ namespace AccelByte.Api
         /// </summary>
         /// <param name="getLikedContentRequest">Detail information for the content request.</param>
         /// <param name="callback">This will be called when the operation succeeded.</param>
+        [Obsolete("This method will be deprecated in future, please use " +
+            "GetLikedContents( GetAllLikedContentRequest getLikedContentRequest ResultCallback<UGCContentsPagingResponse> callback)")]
         public void GetLikedContents(
             GetLikedContentRequest getLikedContentRequest
+            , ResultCallback<UGCContentsPagingResponse> callback)
+        {
+            Report.GetFunctionLog(GetType().Name);
+
+            if (!session.IsValid())
+            {
+                callback.TryError(ErrorCode.IsNotLoggedIn);
+                return;
+            }
+
+            coroutineRunner.Run(
+                api.GetLikedContents(getLikedContentRequest, callback));
+        }
+
+        /// <summary>
+        /// Get Liked Contents.
+        /// </summary>
+        /// <param name="getLikedContentRequest">Detail information for the content request.</param>
+        /// <param name="callback">This will be called when the operation succeeded.</param>
+        public void GetLikedContents(
+            GetAllLikedContentRequest getLikedContentRequest
             , ResultCallback<UGCContentsPagingResponse> callback)
         {
             Report.GetFunctionLog(GetType().Name);