+++ Runtime/Utils/Infoware/PlayStation4Infoware.cs
@@ -1,15 +1,13 @@
 // Copyright (c) 2023 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
-
-using AccelByte.Api;
 using AccelByte.Core;
 using System;
 using UnityEngine;
 
 namespace AccelByte.Utils.Infoware
 {
-    internal class PlayStation4 : InfowareUtils
+    public class PlayStation4Infoware : InfowareUtils
     {
         internal override string GetDeviceID()
         {
@@ -27,25 +25,14 @@ namespace AccelByte.Utils.Infoware
             AccelByteDebug.LogWarning("Incomplete implementation for PlayStation 4 platform");
             return "";
         }
-    }
 
-    internal class PlayStation5 : InfowareUtils
-    {
-        internal override string GetDeviceID()
+        public virtual void GetAccelBytePS4Token(string ps4ClientId, System.Action<string> onGetPlatformTokenFinished)
         {
-            string uniqueIdentifier = SystemInfo.deviceUniqueIdentifier;
-            if (string.IsNullOrEmpty(uniqueIdentifier))
-            {
-                uniqueIdentifier = Guid.NewGuid().ToString();
-            }
-
-            return uniqueIdentifier;
-        }
-
-        internal override string GetMacAddress()
-        {
-            AccelByteDebug.LogWarning("Incomplete implementation for PlayStation 5 platform");
-            return "";
+#if !UNITY_EDITOR && UNITY_PS4
+            throw new NotSupportedException("AccelByte PS4 Extension is not installed. Please contact AccelByte support.");
+#else
+            throw new NotSupportedException("PS4 token generator only available in PS4 package.");
+#endif
         }
     }
 }